requirements.txt
# 镜头检测
transnetv2-pytorch
opencv-python>=4.8.0

# 语音识别
openai-whisper>=20231117

# 说话人检测
resemblyzer>=0.1.4
librosa>=0.10.0
scikit-learn>=1.3.0

# 基础依赖
numpy>=1.24.0
torch>=2.0.0
